refactor preprocessor
