+++ src/preprocessor/tools.cpp
@@ -7,13 +7,14 @@
 #include <boost/numeric/ublas/io.hpp>
 #include <boost/numeric/ublas/lu.hpp>
 #include <boost/numeric/ublas/assignment.hpp>
-#include "Error.h"
+#include "utils/error.h"
 #include "tools.h"
+
 
 using namespace boost::numeric;
 
 
-std::string getElementType(std::string line) {
+std::string preprocessor_tools::getElementType(std::string line) {
     size_t index = line.find(",");
     if (index != std::string::npos) {
         std::string etype = line.substr(index + 1);
@@ -25,7 +26,7 @@ std::string getElementType(std::string line) {
     }
 }
 
-std::unordered_map<std::string, std::optional<double>> getDataFromString(std::string line, std::vector<std::string> keys) {
+std::unordered_map<std::string, std::optional<double>> preprocessor_tools::getDataFromString(std::string line, std::vector<std::string> keys) {
     std::unordered_map<std::string, std::optional<double>> result;
 
     for (const auto& key : keys) {
@@ -61,7 +62,7 @@ std::unordered_map<std::string, std::optional<double>> getDataFromString(std::st
     }
 }
 
-double getSection(std::string line) {
+double preprocessor_tools::getSection(std::string line) {
     size_t index = line.find(",");
     if (index != std::string::npos) {
         std::string section = line.substr(index + 1);
@@ -72,7 +73,7 @@ double getSection(std::string line) {
     }
 }
 
-ublas::matrix<double> TransformMatrix(double angle) {
+ublas::matrix<double> preprocessor_math::TransformMatrix(double angle) {
     ublas::matrix<double> T_matrix(4, 4);
 
     T_matrix <<= cos(angle), sin(angle), 0, 0,
